requirements.txt
langchain-openai>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.2.0
python-dotenv>=1.0.0
langsmith>=0.1.0



